Makes some fixes to gift serializer
