+++ app/controllers/gifts_controller.rb
@@ -1,6 +1,6 @@
 class GiftsController < ApplicationController
   def index
-    gifts = Gift.visible.map{|gift| GiftSerializer.new(gift)}
+    gifts = Gift.visible.map{|gift| Serializers::GiftSerializer.new(gift).to_h}
     render json: gifts
   end
 end
