+++ app/javascript/landing/HomeGrid.js
@@ -1,42 +1,52 @@
-import React from 'react'
-import Button from '@material-ui/core/Button';
-import Card from '@material-ui/core/Card';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import Container from '@material-ui/core/Container';
-import Grid from '@material-ui/core/Grid';
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-
-
+import React, { useEffect } from "react";
+import Button from "@material-ui/core/Button";
+import Card from "@material-ui/core/Card";
+import CardActions from "@material-ui/core/CardActions";
+import CardContent from "@material-ui/core/CardContent";
+import CardMedia from "@material-ui/core/CardMedia";
+import Container from "@material-ui/core/Container";
+import Grid from "@material-ui/core/Grid";
+import { makeStyles } from "@material-ui/core/styles";
+import Typography from "@material-ui/core/Typography";
+import axios from "axios";
 const useStyles = makeStyles((theme) => ({
   cardGrid: {
     paddingTop: theme.spacing(8),
     paddingBottom: theme.spacing(8),
   },
   card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
+    height: "100%",
+    display: "flex",
+    flexDirection: "column",
   },
   cardMedia: {
-    paddingTop: '56.25%', // 16:9
+    paddingTop: "56.25%", // 16:9
   },
   cardContent: {
     flexGrow: 1,
   },
 }));
 
-const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9]; //Replace this with fetch from server
+let gifts = [];
+const fetchGifts = () => {
+  return axios.get("/gifts").then((resp) => {
+    console.log(resp);
+  });
+};
 
 const HomeGrid = () => {
   const classes = useStyles();
-  return(
+
+  useEffect(() => {
+    fetchGifts();
+    return () => {};
+  }, []);
+
+  return (
     <Container className={classes.cardGrid} maxWidth="md">
       <Grid container spacing={4}>
-        {cards.map((card) => (
-          <Grid item key={card} xs={12} sm={6} md={4}>
+        {gifts.map((gift) => (
+          <Grid item key={gift} xs={12} sm={6} md={4}>
             <Card className={classes.card}>
               <CardMedia
                 className={classes.cardMedia}
@@ -46,18 +56,19 @@ const HomeGrid = () => {
               <CardContent className={classes.cardContent}>
                 <Typography gutterBottom variant="h5" component="h2">
                   Heading
-                    </Typography>
+                </Typography>
                 <Typography>
-                  This is a media card. You can use this section to describe the content.
-                    </Typography>
+                  This is a media card. You can use this section to describe the
+                  content.
+                </Typography>
               </CardContent>
               <CardActions>
                 <Button size="small" color="primary">
                   View
-                    </Button>
+                </Button>
                 <Button size="small" color="primary">
                   Edit
-                    </Button>
+                </Button>
               </CardActions>
             </Card>
           </Grid>
@@ -65,6 +76,6 @@ const HomeGrid = () => {
       </Grid>
     </Container>
   );
-}
+};
 
 export default HomeGrid;
